# swagger.yaml
swagger: "2.0"
info:
  title: Toohak API
  description: |
    API description in for Toohak.
    All error strings are required to be clear descriptive message. The descriptive message won't change the marks you receive when being automarked as long as you make an attempt for each one to be custom, but they will improve the usability of the frontend.
  version: 1.0.0
schemes:
  - http
...
components:
  prim:
    Email:
      type: string
      example: hayden.smith@unsw.edu.au
...
  header:
    Token:
      in: header
      name: token
      required: true
      schema:
        type: string
...
  body:
...
    Register:
      in: body
      name: body
      required: true
      schema:
        type: object
        properties:
          email:
            $ref: '#/components/prim/Email'
          password:
            $ref: '#/components/prim/Password'
          nameFirst:
            $ref: '#/components/prim/NameFirst'
          nameLast:
            $ref: '#/components/prim/NameLast'
    Login:
      in: body
      name: body
      required: true
      schema:
        type: object
        properties:
          email:
            $ref: '#/components/prim/Email'
          password:
            $ref: '#/components/prim/Password'
    LogoutIter2:
      in: body
      name: body
      required: true
      schema:
        type: object
        properties:
          token:
            $ref: '#/components/prim/Token'
...
  return:
    Token:
      type: object
      properties:
        token:
          $ref: '#/components/prim/Token'
    Error:
      properties:
        error:
          type: string
          enum:
            - error
    Empty:
      type: object
...
paths:
  /v1/admin/auth/register:
    post:
      summary: Register a new admin user
      description: Takes in information about a new admin user and registers them in the system. This route is not relevant to guests who want to play a particular quiz, but is used for the creation of accounts of people who manage quizzes.
      tags:
        # - "By Topic: Access Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/body/Register'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Token'
        400:
          description: |
            If any of the following are true:
            - Email address is used by another user
            - Email does not satisfy this: https://www.npmjs.com/package/validator (validator.isEmail)
            - NameFirst contains characters other than lowercase letters, uppercase letters, spaces, hyphens, or apostrophes
            - NameFirst is less than 2 characters or more than 20 characters
            - NameLast contains characters other than lowercase letters, uppercase letters, spaces, hyphens, or apostrophes
            - NameLast is less than 2 characters or more than 20 characters
            - Password is less than 8 characters
            - Password does not contain at least one number and at least one letter
          schema:
            $ref: '#/components/return/Error'
  /v1/admin/auth/login:
    post:
      summary: Login an admin user
      description: Takes in information about an admin user to determine if they can log in to manage quizzes. This route is not relevant to guests who want to play a particular quiz, but is used for the creation of accounts of people who manage quizzes. 
      tags:
        # - "By Topic: Access Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/body/Login'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Token'
        400:
          description: |
            If any of the following are true:
            - Email address does not exist
            - Password is not correct for the given email
          schema:
            $ref: '#/components/return/Error'
  /v1/admin/auth/logout:
    post:
      summary: Logs out an admin user who has an active session.
      description: Should be called with a token that is returned after either a login or register has been made.
      tags:
        # - "By Topic: Access Management"
        - "Iteration 2 (New)"
      parameters:
        - $ref: '#/components/body/LogoutIter2'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: Token is not a valid structure
          schema:
            $ref: '#/components/return/Error'
        400:
          description: This token is for a user who has already logged out
          schema:
            $ref: '#/components/return/Error'
  /v2/admin/auth/logout:
    post:
      summary: Logs out an admin user who has an active session.
      description: Should be called with a token that is returned after either a login or register has been made.
      tags:
        # - "By Topic: Access Management"
        - "Iteration 3 (Modified)"
      parameters:
        - $ref: '#/components/header/Token'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: Token is not a valid structure
          schema:
            $ref: '#/components/return/Error'
        400:
          description: This token is for a user who has already logged out
          schema:
            $ref: '#/components/return/Error'
      security:
        - Iteration 3 Token Authentication: []

#  /v1/admin/auth/passwordreset/request:
#    post:
#      summary: Requesting a password reset.
#      description: When users forget their password and can't log in, they can request a security code to begin the process of resetting their password.
#      tags:
#        # - "By Topic: Access Management"
#        - "Iteration 3 (New)"
#      parameters:
#        - $ref: '#/components/body/PasswordResetRequest'
#      responses:
#        200:
#          description: OK
#          schema:
#            $ref: '#/components/return/Empty'
#        400:
#          description: |
#            If any of the following are true:
#            - Email address does not exist
#            Note: This is not the best from a security standpoint, but it's simple for 1531
#          schema:
#            $ref: '#/components/return/Error'
#  /v1/admin/auth/passwordreset/reset:
#    post:
#      summary: Completing a password reset.
#      description: When users forget their password and get a security code they can complete the password reset.
#      tags:
#        # - "By Topic: Access Management"
#        - "Iteration 3 (New)"
#      parameters:
#        - $ref: '#/components/body/PasswordResetReset'
#      responses:
#        200:
#          description: OK
#          schema:
#            $ref: '#/components/return/Empty'
#        400:
#          description: |
#            If any of the following are true:
#            - Reset code is not valid
#            - New Password is less than 8 characters
#            - New Password does not contain at least one number and at least one letter
#            - Password has already been used by this user before
#          schema:
#            $ref: '#/components/return/Error'
  /v1/admin/user/details:
    get:
      summary: Get the details of an admin user.
      description: For the given admin user that is logged in, return all of the relevant details.
      tags:
        # - "By Topic: Access Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/body/UserDetailsGetIter2'
      responses:
        200:
          description: |
            - "name" is the first and last name concatenated with a single space between them
            - numSuccessfulLogins includes logins direct via registration, and is counted from the moment of registration starting at 1
            - numFailedPasswordsSinceLastLogin is reset every time they have a successful login, and simply counts the number of attempted logins that failed due to incorrect password, only since the last login
          schema:
            $ref: '#/components/return/User'
        401:
          description: Token is not a valid structure
          schema:
            $ref: '#/components/return/Error'
        403:
          description: Provided token is valid structure, but is not for a currently logged in session
          schema:
            $ref: '#/components/return/Error'
    put:
      summary: Update the details of an admin user (non-password).
      description: Given a set of properties, update those properties of this logged in admin user.
      tags:
        # - "By Topic: Access Management"
        - "Iteration 2 (New)"
      parameters:
        - $ref: '#/components/body/UserDetailsPutIter2'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: Token is not a valid structure
          schema:
            $ref: '#/components/return/Error'
        403:
          description: Provided token is valid structure, but is not for a currently logged in session
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - Email is currently used by another user (excluding the current authorised user)
            - Email does not satisfy this: https://www.npmjs.com/package/validator (validator.isEmail)
            - NameFirst contains characters other than lowercase letters, uppercase letters, spaces, hyphens, or apostrophes
            - NameFirst is less than 2 characters or more than 20 characters
            - NameLast contains characters other than lowercase letters, uppercase letters, spaces, hyphens, or apostrophes
            - NameLast is less than 2 characters or more than 20 characters
          schema:
            $ref: '#/components/return/Error'
  /v2/admin/user/details:
    get:
      summary: Get the details of an admin user.
      description: For the given admin user that is logged in, return all of the relevant details.
      tags:
        # - "By Topic: Access Management"
        - "Iteration 3 (Modified)"
      parameters:
        - $ref: '#/components/header/Token'
      responses:
        200:
          description: |
            - "name" is the first and last name concatenated with a single space between them
            - numSuccessfulLogins includes logins direct via registration, and is counted from the moment of registration starting at 1
            - numFailedPasswordsSinceLastLogin is reset every time they have a successful login, and simply counts the number of attempted logins that failed due to incorrect password, only since the last login
          schema:
            $ref: '#/components/return/User'
        401:
          description: Token is not a valid structure
          schema:
            $ref: '#/components/return/Error'
        403:
          description: Provided token is valid structure, but is not for a currently logged in session
          schema:
            $ref: '#/components/return/Error'
      security:
        - Iteration 3 Token Authentication: []
    put:
      summary: Update the details of an admin user (non-password).
      description: Given a set of properties, update those properties of this logged in admin user.
      tags:
        # - "By Topic: Access Management"
        - "Iteration 3 (Modified)"
      parameters:
        - $ref: '#/components/header/Token'
        - $ref: '#/components/body/UserDetailsPutIter3'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: Token is not a valid structure
          schema:
            $ref: '#/components/return/Error'
        403:
          description: Provided token is valid structure, but is not for a currently logged in session
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - Email is currently used by another user (excluding the current authorised user)
            - Email does not satisfy this: https://www.npmjs.com/package/validator (validator.isEmail)
            - NameFirst contains characters other than lowercase letters, uppercase letters, spaces, hyphens, or apostrophes
            - NameFirst is less than 2 characters or more than 20 characters
            - NameLast contains characters other than lowercase letters, uppercase letters, spaces, hyphens, or apostrophes
            - NameLast is less than 2 characters or more than 20 characters
          schema:
            $ref: '#/components/return/Error'
      security:
        - Iteration 3 Token Authentication: []
  /v1/admin/user/password:
    put:
      summary: Update the password of an admin user.
      description: Given details relating to a password change, update the password of a logged in user.
      tags:
        # - "By Topic: Access Management"
        - "Iteration 2 (New)"
      parameters:
        - $ref: '#/components/body/UserPasswordIter2'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: Token is not a valid structure
          schema:
            $ref: '#/components/return/Error'
        403:
          description: Provided token is valid structure, but is not for a currently logged in session
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - Old Password is not the correct old password
            - Old Password and New Password match exactly
            - New Password has already been used before by this user
            - New Password is less than 8 characters
            - New Password does not contain at least one number and at least one letter
          schema:
            $ref: '#/components/return/Error'
  /v2/admin/user/password:
    put:
      summary: Update the password of an admin user.
      description: Given details relating to a password change, update the password of a logged in user.
      tags:
        # - "By Topic: Access Management"
        - "Iteration 3 (Modified)"
      parameters:
        - $ref: '#/components/header/Token'
        - $ref: '#/components/body/UserPasswordIter3'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: Token is not a valid structure
          schema:
            $ref: '#/components/return/Error'
        403:
          description: Provided token is valid structure, but is not for a currently logged in session
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - Old Password is not the correct old password
            - Old Password and New Password match exactly
            - New Password has already been used before by this user
            - New Password is less than 8 characters
            - New Password does not contain at least one number and at least one letter
          schema:
            $ref: '#/components/return/Error'
      security:
        - Iteration 3 Token Authentication: []
  /v1/admin/quiz/list:
    get:
      summary: Lists all user's quizzes
      description: Provide a list of all quizzes that are owned by the currently logged in user
      tags:
        # - "By Topic: Quiz Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/query/Token'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/QuizList'
        401:
          description: Token is not a valid structure
          schema:
            $ref: '#/components/return/Error'
        403:
          description: Provided token is valid structure, but is not for a currently logged in session
          schema:
            $ref: '#/components/return/Error'
  /v2/admin/quiz/list:
    get:
      summary: Lists all user's quizzes
      description: Provide a list of all quizzes that are owned by the currently logged in user
      tags:
        # - "By Topic: Quiz Management"
        - "Iteration 3 (Modified)"
      parameters:
        - $ref: '#/components/header/Token'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/QuizList'
        401:
          description: Token is not a valid structure
          schema:
            $ref: '#/components/return/Error'
        403:
          description: Provided token is valid structure, but is not for a currently logged in session
          schema:
            $ref: '#/components/return/Error'
      security:
        - Iteration 3 Token Authentication: []
  /v1/admin/quiz:
    post:
      summary: Create a new quiz
      description: Given basic details about a new quiz, create one for the logged in user
      tags:
        # - "By Topic: Quiz Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/body/QuizNewIter2'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/QuizNew'
        401:
          description: Token is not a valid structure
          schema:
            $ref: '#/components/return/Error'
        403:
          description: Provided token is valid structure, but is not for a currently logged in session
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - Name contains invalid characters. Valid characters are alphanumeric and spaces
            - Name is either less than 3 characters long or more than 30 characters long
            - Name is already used by the current logged in user for another quiz
            - Description is more than 100 characters in length (note: empty strings are OK)
          schema:
            $ref: '#/components/return/Error'
  /v2/admin/quiz:
    post:
      summary: Create a new quiz
      description: Given basic details about a new quiz, create one for the logged in user
      tags:
        # - "By Topic: Quiz Management"
        - "Iteration 3 (Modified)"
      parameters:
        - $ref: '#/components/header/Token'
        - $ref: '#/components/body/QuizNewIter3'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/QuizNew'
        401:
          description: Token is not a valid structure
          schema:
            $ref: '#/components/return/Error'
        403:
          description: Provided token is valid structure, but is not for a currently logged in session
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - Name contains invalid characters. Valid characters are alphanumeric and spaces
            - Name is either less than 3 characters long or more than 30 characters long
            - Name is already used by the current logged in user for another quiz
            - Description is more than 100 characters in length (note: empty strings are OK)
          schema:
            $ref: '#/components/return/Error'
      security:
        - Iteration 3 Token Authentication: []
...
  /v1/admin/quiz/{quizid}/name:
    put:
      summary: Update quiz name
      description: Update the name of the relevant quiz
      tags:
        # - "By Topic: Quiz Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/path/QuizId'
        - $ref: '#/components/body/QuizUpdateNameIter2'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: Token is not a valid structure
          schema:
            $ref: '#/components/return/Error'
        403:
          description: Provided token is valid structure, but is not for a currently logged in session
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - Quiz ID does not refer to a valid quiz
            - Quiz ID does not refer to a quiz that this user owns
            - Name contains invalid characters. Valid characters are alphanumeric and spaces
            - Name is either less than 3 characters long or more than 30 characters long
            - Name is already used by the current logged in user for another quiz
          schema:
            $ref: '#/components/return/Error'
...
  /v1/admin/quiz/trash:
    get:
      summary: View the quizzes in trash
      description: View the quizzes that are currently in the trash for the logged in user
      tags:
        # - "By Topic: Quiz Management"
        - "Iteration 2 (New)"
      parameters:
        - $ref: '#/components/body/QuizTrash'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/QuizTrash'
        401:
          description: Token is not a valid structure
          schema:
            $ref: '#/components/return/Error'
        403:
          description: Provided token is valid structure, but is not for a currently logged in session
          schema:
            $ref: '#/components/return/Error'
...
securityDefinitions:
  Iteration 3 Token Authentication:
    type: apiKey
    name: token
    in: header